+++ src/App.tsx
@@ -16,7 +16,7 @@ import { useLocation } from 'react-router-dom';
 import ReactGA from "react-ga4";
 import { CreatorViewMode } from './components/creator/Editor/CreatorViewMode';
 import { useThemeContext } from './theme/ThemeContext';
-import { BlocklyPrueba } from './components/BlocklyPrueba';
+import { ToolboxPreview } from './components/creator/Editor/ChallengeDetailsEdition/ToolboxPreview';
 
 const AnalyticsComponent = () => {
   const location = useLocation();
@@ -84,7 +84,7 @@ const router = createHashRouter([{
   },
   {
     path: "/blocklyPrueba",
-    element: <BlocklyPrueba/>
+    element: <ToolboxPreview/>
   }
 ]}]);
